requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
discord.py==1.3.2
dnspython==1.16.0
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
mysql-connector-python==8.0.19
protobuf==3.6.1
pycparser==2.19
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.3
websockets==8.1
Werkzeug==1.0.0
yarl==1.4.2
